+++ src/rwlexpreval.c
@@ -14,6 +14,7 @@
  *
  * History
  *
+ * bengsig  13-aug-2021 - Add break
  * bengsig  09-aug-2021 - Use constants rwl_xxxxp
  * bengsig  09-aug-2021 - Fix error calling proc with db from main
  * bengsig  22-jun-2021 - Add epochseconds
@@ -1210,19 +1211,6 @@ void rwlexpreval ( rwl_estack *stk , rwl_location *loc , rwl_xeqenv *xev , rwl_v
         {
 	  ub4 pp;
 	  vv = &xev->evar[stk[i].esvar];
-
-	  // If we attempt calling a function that includes SQL
-	  // we must have a database
-	  if ( vv
-	     && RWL_STACK_FUNCCALL == stk[i].elemtype
-	     && RWL_CODE_SQLHEAD == xev->rwm->code[vv->vval].ctyp
-	     && !xev->curdb->username)
-	  {
-	    rwlexecerror(xev, loc, RWL_ERROR_NOT_DONE_IN_MAIN, 
-	      "function with database execution");
-	    rwlcopyvalue(cstak+i, rwl_nullp);
-	    goto pop_N;
-	  }
 
 	  // If random procedure
 	  // pick a random one of them